requirements copy.txt
boto==2.49.0
boto3==1.9.130
botocore==1.12.130
Collectfast==0.6.2
dj-database-url==0.5.0
dj-static==0.0.6
Django==2.2
django-appconf==1.0.3
django-boto==0.3.12
django-ckeditor==5.6.1
django-compressor==2.2
django-debug-toolbar==1.11
django-environ==0.4.5
django-extra-views==0.12.0
django-js-asset==1.2.2
django-markdown-deux==1.0.5
django-markdownx==2.0.28
django-storages==1.7.1
django-toolbelt==0.0.1
docutils==0.14
gunicorn==19.9.0
jmespath==0.9.4
Markdown==3.1
markdown2==2.3.7
Pillow==6.0.0
python-dateutil==2.8.0
pytz==2018.9
rcssmin==1.0.6
rjsmin==1.0.12
s3transfer==0.2.0
six==1.12.0
sqlparse==0.3.0
static3==0.7.0
urllib3==1.24.1
